requirements.txt
flask
flask_sqlalchemy
flask_jwt
functools
werkzeug.security